// src/frontend/src/app/api/reports/swagger.ts
export enum Outcome {
    Unknow = 0,
    Planed = 1,
    Passed = 2,
    Failed = 3,
    Blocked = 4,
    Skipped = 5,
    Paused = 6,
}

export class SetTesterRequest implements ISetTesterRequest {
    testPointIds?: number[] | undefined;
    testerId!: number;

    constructor(data?: ISetTesterRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["testPointIds"])) {
                this.testPointIds = [] as any;
                for (let item of _data["testPointIds"])
                    this.testPointIds!.push(item);
            }
            this.testerId = _data["testerId"];
        }
    }

    static fromJS(data: any): SetTesterRequest {
        data = typeof data === 'object' ? data : {};
        let result = new SetTesterRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.testPointIds)) {
            data["testPointIds"] = [];
            for (let item of this.testPointIds)
                data["testPointIds"].push(item);
        }
        data["testerId"] = this.testerId;
        return data;
    }
}

export interface ISetTesterRequest {
    testPointIds?: number[] | undefined;
    testerId: number;
}

export class SetOutcomeRequest implements ISetOutcomeRequest {
    testPointIds?: number[] | undefined;
    outcome!: Outcome;

    constructor(data?: ISetOutcomeRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["testPointIds"])) {
                this.testPointIds = [] as any;
                for (let item of _data["testPointIds"])
                    this.testPointIds!.push(item);
            }
            this.outcome = _data["outcome"];
        }
    }

    static fromJS(data: any): SetOutcomeRequest {
        data = typeof data === 'object' ? data : {};
        let result = new SetOutcomeRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.testPointIds)) {
            data["testPointIds"] = [];
            for (let item of this.testPointIds)
                data["testPointIds"].push(item);
        }
        data["outcome"] = this.outcome;
        return data;
    }
}

export interface ISetOutcomeRequest {
    testPointIds?: number[] | undefined;
    outcome: Outcome;
}

export class SetDescriptionRequest implements ISetDescriptionRequest {
    testPointIds?: number[] | undefined;
    description?: string | undefined;

    constructor(data?: ISetDescriptionRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["testPointIds"])) {
                this.testPointIds = [] as any;
                for (let item of _data["testPointIds"])
                    this.testPointIds!.push(item);
            }
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): SetDescriptionRequest {
        data = typeof data === 'object' ? data : {};
        let result = new SetDescriptionRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.testPointIds)) {
            data["testPointIds"] = [];
            for (let item of this.testPointIds)
                data["testPointIds"].push(item);
        }
        data["description"] = this.description;
        return data;
    }
}

export interface ISetDescriptionRequest {
    testPointIds?: number[] | undefined;
    description?: string | undefined;
}
// ...
export class AddTestCaseToSuiteRequest implements IAddTestCaseToSuiteRequest {
    testCasesIds?: number[] | undefined;

    constructor(data?: IAddTestCaseToSuiteRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["testCasesIds"])) {
                this.testCasesIds = [] as any;
                for (let item of _data["testCasesIds"])
                    this.testCasesIds!.push(item);
            }
        }
    }

    static fromJS(data: any): AddTestCaseToSuiteRequest {
        data = typeof data === 'object' ? data : {};
        let result = new AddTestCaseToSuiteRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.testCasesIds)) {
            data["testCasesIds"] = [];
            for (let item of this.testCasesIds)
                data["testCasesIds"].push(item);
        }
        return data;
    }
}

export interface IAddTestCaseToSuiteRequest {
    testCasesIds?: number[] | undefined;
}

export class DeleteTestCaseFromSuiteRequest implements IDeleteTestCaseFromSuiteRequest {
    suiteTestCasesIds?: number[] | undefined;

    constructor(data?: IDeleteTestCaseFromSuiteRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["suiteTestCasesIds"])) {
                this.suiteTestCasesIds = [] as any;
                for (let item of _data["suiteTestCasesIds"])
                    this.suiteTestCasesIds!.push(item);
            }
        }
    }

    static fromJS(data: any): DeleteTestCaseFromSuiteRequest {
        data = typeof data === 'object' ? data : {};
        let result = new DeleteTestCaseFromSuiteRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.suiteTestCasesIds)) {
            data["suiteTestCasesIds"] = [];
            for (let item of this.suiteTestCasesIds)
                data["suiteTestCasesIds"].push(item);
        }
        return data;
    }
}

export interface IDeleteTestCaseFromSuiteRequest {
    suiteTestCasesIds?: number[] | undefined;
}
// ...
export class SetTestCaseConfigRequest implements ISetTestCaseConfigRequest {
    suiteTestCaseIds?: number[] | undefined;
    testConfigIds?: number[] | undefined;

    constructor(data?: ISetTestCaseConfigRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["suiteTestCaseIds"])) {
                this.suiteTestCaseIds = [] as any;
                for (let item of _data["suiteTestCaseIds"])
                    this.suiteTestCaseIds!.push(item);
            }
            if (Array.isArray(_data["testConfigIds"])) {
                this.testConfigIds = [] as any;
                for (let item of _data["testConfigIds"])
                    this.testConfigIds!.push(item);
            }
        }
    }

    static fromJS(data: any): SetTestCaseConfigRequest {
        data = typeof data === 'object' ? data : {};
        let result = new SetTestCaseConfigRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.suiteTestCaseIds)) {
            data["suiteTestCaseIds"] = [];
            for (let item of this.suiteTestCaseIds)
                data["suiteTestCaseIds"].push(item);
        }
        if (Array.isArray(this.testConfigIds)) {
            data["testConfigIds"] = [];
            for (let item of this.testConfigIds)
                data["testConfigIds"].push(item);
        }
        return data;
    }
}

export interface ISetTestCaseConfigRequest {
    suiteTestCaseIds?: number[] | undefined;
    testConfigIds?: number[] | undefined;
}
// ...
export class SetTestSuiteConfigRequest implements ISetTestSuiteConfigRequest {
    testConfigIds?: number[] | undefined;

    constructor(data?: ISetTestSuiteConfigRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["testConfigIds"])) {
                this.testConfigIds = [] as any;
                for (let item of _data["testConfigIds"])
                    this.testConfigIds!.push(item);
            }
        }
    }

    static fromJS(data: any): SetTestSuiteConfigRequest {
        data = typeof data === 'object' ? data : {};
        let result = new SetTestSuiteConfigRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.testConfigIds)) {
            data["testConfigIds"] = [];
            for (let item of this.testConfigIds)
                data["testConfigIds"].push(item);
        }
        return data;
    }
}

export interface ISetTestSuiteConfigRequest {
    testConfigIds?: number[] | undefined;
}

export class AddOrUpdateSuiteRequest implements IAddOrUpdateSuiteRequest {
    parentId!: number;
    name?: string | undefined;

    constructor(data?: IAddOrUpdateSuiteRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.parentId = _data["parentId"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): AddOrUpdateSuiteRequest {
        data = typeof data === 'object' ? data : {};
        let result = new AddOrUpdateSuiteRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["parentId"] = this.parentId;
        data["name"] = this.name;
        return data;
    }
}

export interface IAddOrUpdateSuiteRequest {
    parentId: number;
    name?: string | undefined;
}